+++ src/vs/editor/contrib/inlineCompletions/browser/model/inlineSuggestionItem.ts
@@ -4,7 +4,6 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { BugIndicatingError } from '../../../../../base/common/errors.js';
-import { matchesSubString } from '../../../../../base/common/filters.js';
 import { IObservable, ITransaction, observableSignal, observableValue } from '../../../../../base/common/observable.js';
 import { commonPrefixLength, commonSuffixLength, splitLines } from '../../../../../base/common/strings.js';
 import { URI } from '../../../../../base/common/uri.js';
@@ -21,12 +20,13 @@ import { PositionOffsetTransformerBase } from '../../../../common/core/text/posi
 import { TextLength } from '../../../../common/core/text/textLength.js';
 import { linesDiffComputers } from '../../../../common/diff/linesDiffComputers.js';
 import { Command, IInlineCompletionHint, InlineCompletion, InlineCompletionEndOfLifeReason, InlineCompletionHintStyle, InlineCompletionTriggerKind, InlineCompletionWarning, PartialAcceptInfo } from '../../../../common/languages.js';
-import { EndOfLinePreference, ITextModel } from '../../../../common/model.js';
+import { ITextModel } from '../../../../common/model.js';
 import { TextModelText } from '../../../../common/model/textModelText.js';
 import { InlineCompletionViewData, InlineCompletionViewKind } from '../view/inlineEdits/inlineEditsViewInterface.js';
 import { computeEditKind, InlineSuggestionEditKind } from './editKind.js';
-import { IInlineSuggestDataAction, IInlineSuggestDataActionEdit, InlineSuggestAlternativeAction, InlineSuggestData, InlineSuggestionList, PartialAcceptance, RenameInfo, SnippetInfo } from './provideInlineCompletions.js';
-import { singleTextRemoveCommonPrefix } from './singleTextEditHelpers.js';
+import { inlineCompletionIsVisible } from './inlineCompletionIsVisible.js';
+import { IInlineSuggestDataAction, IInlineSuggestDataActionEdit, InlineSuggestData, InlineSuggestionList, PartialAcceptance, RenameInfo, SnippetInfo } from './provideInlineCompletions.js';
+import { InlineSuggestAlternativeAction } from './InlineSuggestAlternativeAction.js';
 
 export type InlineSuggestionItem = InlineEditItem | InlineCompletionItem;
 
@@ -365,46 +365,6 @@ export class InlineCompletionItem extends InlineSuggestionItemBase {
 
 	public get editRange(): Range { return this.getSingleTextEdit().range; }
 	public get insertText(): string { return this.getSingleTextEdit().text; }
-}
-
-export function inlineCompletionIsVisible(singleTextEdit: TextReplacement, originalRange: Range | undefined, model: ITextModel, cursorPosition: Position): boolean {
-	const minimizedReplacement = singleTextRemoveCommonPrefix(singleTextEdit, model);
-	const editRange = singleTextEdit.range;
-	if (!editRange
-		|| (originalRange && !originalRange.getStartPosition().equals(editRange.getStartPosition()))
-		|| cursorPosition.lineNumber !== minimizedReplacement.range.startLineNumber
-		|| minimizedReplacement.isEmpty // if the completion is empty after removing the common prefix of the completion and the model, the completion item would not be visible
-	) {
-		return false;
-	}
-
-	// We might consider comparing by .toLowerText, but this requires GhostTextReplacement
-	const originalValue = model.getValueInRange(minimizedReplacement.range, EndOfLinePreference.LF);
-	const filterText = minimizedReplacement.text;
-
-	const cursorPosIndex = Math.max(0, cursorPosition.column - minimizedReplacement.range.startColumn);
-
-	let filterTextBefore = filterText.substring(0, cursorPosIndex);
-	let filterTextAfter = filterText.substring(cursorPosIndex);
-
-	let originalValueBefore = originalValue.substring(0, cursorPosIndex);
-	let originalValueAfter = originalValue.substring(cursorPosIndex);
-
-	const originalValueIndent = model.getLineIndentColumn(minimizedReplacement.range.startLineNumber);
-	if (minimizedReplacement.range.startColumn <= originalValueIndent) {
-		// Remove indentation
-		originalValueBefore = originalValueBefore.trimStart();
-		if (originalValueBefore.length === 0) {
-			originalValueAfter = originalValueAfter.trimStart();
-		}
-		filterTextBefore = filterTextBefore.trimStart();
-		if (filterTextBefore.length === 0) {
-			filterTextAfter = filterTextAfter.trimStart();
-		}
-	}
-
-	return filterTextBefore.startsWith(originalValueBefore)
-		&& !!matchesSubString(originalValueAfter, filterTextAfter);
 }
 
 export class InlineEditItem extends InlineSuggestionItemBase {
